Fixed some state problems, a memory leak and hopefully a floating point error but really I don't know what I'm looking for as of yet.

+++ Classes/Instructions/Arithmetic/VCInstructionDivideDelegate.m
@@ -51,11 +51,12 @@
 	int finalValue = 0;
 	
 	if(valueOne != 0) {
-		finalValue = valueZero / valueOne;		
+		finalValue = (int)((float)valueZero / (float)valueOne);		
 	} else {
 		finalValue = 0;
 	}
-
+	
+	
 	
 	[registers setValueOfRegister:fieldTwo withValue:finalValue];
